+++ test/truffle/compiler/pe.sh
@@ -2,4 +2,4 @@
 
 source test/truffle/common.sh.inc
 
-jt ruby --graal -J-Dgraal.TruffleCompilationExceptionsAreThrown=true -J-Dgraal.TruffleIterativePartialEscape=true -Xbasic_ops.inline=false test/truffle/compiler/pe/pe.rb "$@"
+jt ruby --graal --jvm.Dgraal.TruffleCompilationExceptionsAreThrown=true --jvm.Dgraal.TruffleIterativePartialEscape=true -Xbasic_ops.inline=false test/truffle/compiler/pe/pe.rb "$@"